refactor: add test files for prayer requests, including fixtures and permission tests
