+++ backend/tests/prayer_requests/test_api.py
@@ -1,161 +1,10 @@
-from datetime import date
 from unittest.mock import patch
 
 import pytest
 from django.urls import reverse
-from django.utils import timezone
 from rest_framework import status
 
 from apps.prayer_requests.models import PrayerRequest, PrayerRequestFollowUp
-
-
-@pytest.fixture
-def pastor_user(create_user):
-    """Create a pastor user."""
-    return create_user(
-        username="pastor",
-        email="pastor@example.com",
-        role="pastor",
-        is_staff=True,
-    )
-
-
-@pytest.fixture
-def pastor_client(api_client, pastor_user):
-    """Return an authenticated API client for pastor user."""
-    from rest_framework_simplejwt.tokens import RefreshToken
-
-    refresh = RefreshToken.for_user(pastor_user)
-    api_client.credentials(HTTP_AUTHORIZATION=f"Bearer {refresh.access_token}")
-    return api_client
-
-
-@pytest.fixture
-def member_profile(db, user):
-    """Create a member profile for the test user."""
-    from apps.members.models import Member
-
-    return Member.objects.create(
-        user=user,
-        first_name="Test",
-        last_name="Member",
-        email="testmember@example.com",
-        phone="1234567890",
-        date_of_birth=date(1990, 1, 1),
-        is_active=True,
-    )
-
-
-@pytest.fixture
-def prayer_request(db, member_profile):
-    """Create a basic prayer request."""
-    return PrayerRequest.objects.create(
-        title="Test Prayer Request",
-        description="Please pray for my health.",
-        category="health",
-        requester=member_profile,
-        status="pending",
-        priority="medium",
-    )
-
-
-@pytest.fixture
-def assigned_prayer_request(db, member_profile, pastor_user):
-    """Create an assigned prayer request."""
-    return PrayerRequest.objects.create(
-        title="Assigned Prayer Request",
-        description="This request has been assigned.",
-        category="family",
-        requester=member_profile,
-        status="assigned",
-        priority="high",
-        assigned_to=pastor_user,
-        assigned_at=timezone.now(),
-    )
-
-
-@pytest.fixture
-def public_prayer_request(db):
-    """Create a public prayer request."""
-    return PrayerRequest.objects.create(
-        title="Public Prayer Request",
-        description="Please pray for our community.",
-        category="other",
-        is_public=True,
-        requester_name="Community Member",
-        status="pending",
-    )
-
-
-@pytest.fixture
-def follow_up(db, prayer_request, pastor_user):
-    """Create a follow-up for a prayer request."""
-    return PrayerRequestFollowUp.objects.create(
-        prayer_request=prayer_request,
-        action_type="prayed",
-        notes="Prayed for this request during service.",
-        created_by=pastor_user,
-    )
-
-
-# =============================================================================
-# Model Tests
-# =============================================================================
-@pytest.mark.django_db
-class TestPrayerRequestModel:
-    """Tests for the PrayerRequest model."""
-
-    def test_prayer_request_str(self, prayer_request):
-        """Test string representation."""
-        assert "Test Prayer Request" in str(prayer_request)
-        assert "Pending" in str(prayer_request)
-
-    def test_requester_display_name_with_member(self, prayer_request):
-        """Test requester_display_name with member requester."""
-        assert prayer_request.requester_display_name == "Test Member"
-
-    def test_requester_display_name_anonymous(self, prayer_request):
-        """Test requester_display_name when anonymous."""
-        prayer_request.is_anonymous = True
-        prayer_request.save()
-        assert prayer_request.requester_display_name == "Anonymous"
-
-    def test_requester_display_name_non_member(self, db):
-        """Test requester_display_name for non-member submission."""
-        pr = PrayerRequest.objects.create(
-            title="Non-member Request",
-            description="Test",
-            requester_name="John Visitor",
-        )
-        assert pr.requester_display_name == "John Visitor"
-
-    def test_assign_to(self, prayer_request, pastor_user, admin_user):
-        """Test assigning a prayer request."""
-        prayer_request.assign_to(pastor_user, assigned_by=admin_user)
-
-        prayer_request.refresh_from_db()
-        assert prayer_request.assigned_to == pastor_user
-        assert prayer_request.assigned_by == admin_user
-        assert prayer_request.assigned_at is not None
-        assert prayer_request.status == "assigned"
-
-    def test_mark_completed(self, prayer_request):
-        """Test marking a prayer request as completed."""
-        prayer_request.mark_completed()
-
-        prayer_request.refresh_from_db()
-        assert prayer_request.status == "completed"
-        assert prayer_request.completed_at is not None
-
-
-@pytest.mark.django_db
-class TestPrayerRequestFollowUpModel:
-    """Tests for the PrayerRequestFollowUp model."""
-
-    def test_follow_up_str(self, follow_up):
-        """Test string representation."""
-        assert "Prayed For" in str(follow_up)
-        assert "Test Prayer Request" in str(follow_up)
 
 
 # =============================================================================
@@ -401,43 +250,6 @@ class TestPrayerRequestCustomEndpoints:
         assert response.data["total"] == 2
         assert response.data["pending"] == 1
         assert response.data["assigned"] == 1
-
-
-# =============================================================================
-# Service Tests
-# =============================================================================
-@pytest.mark.django_db
-class TestPrayerRequestServices:
-    """Tests for prayer request services."""
-
-    @patch("apps.prayer_requests.services.send_mail")
-    def test_notify_assignment_sends_email(self, mock_send_mail, assigned_prayer_request):
-        """Test that notify_assignment sends an email."""
-        from apps.prayer_requests.services import notify_assignment
-
-        result = notify_assignment(assigned_prayer_request)
-
-        assert result is True
-        mock_send_mail.assert_called_once()
-
-    def test_notify_assignment_no_assignee(self, prayer_request):
-        """Test notify_assignment when no assigned user."""
-        from apps.prayer_requests.services import notify_assignment
-
-        result = notify_assignment(prayer_request)
-        assert result is False
-
-    def test_get_statistics(self, prayer_request, assigned_prayer_request):
-        """Test get_prayer_request_statistics."""
-        from apps.prayer_requests.services import get_prayer_request_statistics
-
-        stats = get_prayer_request_statistics()
-
-        assert stats["total"] == 2
-        assert stats["pending"] == 1
-        assert stats["assigned"] == 1
-        assert "by_category" in stats
-        assert "by_priority" in stats
 
 
 # =============================================================================
